SW-1749 Break up large test classes (part 1)
- AccessionModelTest.kt
- Divide the tests into several categories and place in new files
- Place shared code into a Helpers.kt file
