+++ src/test/kotlin/com/terraformation/backend/seedbank/model/accession/AccessionModelCalculationsTest.kt
@@ -1,870 +1,28 @@
-package com.terraformation.backend.seedbank.model
+package com.terraformation.backend.seedbank.model.accession
 
-import com.terraformation.backend.assertJsonEquals
-import com.terraformation.backend.db.default_schema.UserId
 import com.terraformation.backend.db.seedbank.AccessionId
 import com.terraformation.backend.db.seedbank.AccessionState
-import com.terraformation.backend.db.seedbank.DataSource
 import com.terraformation.backend.db.seedbank.ProcessingMethod
 import com.terraformation.backend.db.seedbank.SeedQuantityUnits
-import com.terraformation.backend.db.seedbank.ViabilityTestId
-import com.terraformation.backend.db.seedbank.ViabilityTestResultId
-import com.terraformation.backend.db.seedbank.ViabilityTestSubstrate
-import com.terraformation.backend.db.seedbank.ViabilityTestType
-import com.terraformation.backend.db.seedbank.WithdrawalId
 import com.terraformation.backend.db.seedbank.WithdrawalPurpose
 import com.terraformation.backend.seedbank.grams
 import com.terraformation.backend.seedbank.milligrams
+import com.terraformation.backend.seedbank.model.AccessionModel
+import com.terraformation.backend.seedbank.model.SeedQuantityModel
+import com.terraformation.backend.seedbank.model.WithdrawalModel
 import com.terraformation.backend.seedbank.seeds
 import java.math.BigDecimal
-import java.time.Clock
 import java.time.Instant
-import java.time.LocalDate
-import java.time.ZoneOffset
 import org.junit.jupiter.api.Assertions.assertEquals
-import org.junit.jupiter.api.Assertions.assertNotEquals
 import org.junit.jupiter.api.Assertions.assertNull
 import org.junit.jupiter.api.DynamicTest
 import org.junit.jupiter.api.Nested
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.TestFactory
 import org.junit.jupiter.api.assertAll
-import org.junit.jupiter.api.assertDoesNotThrow
 import org.junit.jupiter.api.assertThrows
 
-internal class AccessionModelTest {
-  private val today = january(2)
-  private val todayInstant = today.atStartOfDay(ZoneOffset.UTC).toInstant()
-  private val clock: Clock = Clock.fixed(todayInstant, ZoneOffset.UTC)
-  private val tomorrow = today.plusDays(1)
-  private val tomorrowInstant = tomorrow.atStartOfDay(ZoneOffset.UTC).toInstant()
-  private val tomorrowClock = Clock.fixed(tomorrowInstant, ZoneOffset.UTC)
-  private val yesterday = today.minusDays(1)
-  private val yesterdayInstant = yesterday.atStartOfDay(ZoneOffset.UTC).toInstant()
-  private val yesterdayClock = Clock.fixed(yesterdayInstant, ZoneOffset.UTC)
-
-  private var viabilityTestResultId = ViabilityTestResultId(1)
-  private var viabilityTestId = ViabilityTestId(1)
-  private var withdrawalId = WithdrawalId(1)
-  private var defaultState = AccessionState.Processing
-
-  private fun accession(
-      viabilityTests: List<ViabilityTestModel> = emptyList(),
-      cutTestSeedsCompromised: Int? = null,
-      cutTestSeedsEmpty: Int? = null,
-      cutTestSeedsFilled: Int? = null,
-      dryingEndDate: LocalDate? = null,
-      dryingStartDate: LocalDate? = null,
-      processingStartDate: LocalDate? = null,
-      subsetCount: Int? = null,
-      state: AccessionState = defaultState,
-      storageLocation: String? = null,
-      storagePackets: Int? = null,
-      storageStartDate: LocalDate? = null,
-      subsetWeight: SeedQuantityModel? = null,
-      total: SeedQuantityModel? = null,
-      withdrawals: List<WithdrawalModel> = emptyList(),
-      processingMethod: ProcessingMethod? =
-          total?.let { if (it.grams != null) ProcessingMethod.Weight else ProcessingMethod.Count },
-  ): AccessionModel {
-    return AccessionModel(
-        id = AccessionId(1L),
-        accessionNumber = "dummy",
-        createdTime = clock.instant(),
-        cutTestSeedsCompromised = cutTestSeedsCompromised,
-        cutTestSeedsEmpty = cutTestSeedsEmpty,
-        cutTestSeedsFilled = cutTestSeedsFilled,
-        dryingEndDate = dryingEndDate,
-        dryingStartDate = dryingStartDate,
-        processingMethod = processingMethod,
-        processingStartDate = processingStartDate,
-        source = DataSource.Web,
-        state = state,
-        storageLocation = storageLocation,
-        storagePackets = storagePackets,
-        storageStartDate = storageStartDate,
-        subsetCount = subsetCount,
-        subsetWeightQuantity = subsetWeight,
-        total = total,
-        viabilityTests = viabilityTests,
-        withdrawals = withdrawals,
-    )
-  }
-
-  private fun nextViabilityTestId(): ViabilityTestId {
-    val current = viabilityTestId
-    viabilityTestId = ViabilityTestId(current.value + 1)
-    return current
-  }
-
-  private fun viabilityTest(
-      testType: ViabilityTestType = ViabilityTestType.Lab,
-      startDate: LocalDate? = january(1),
-      seedsTested: Int? = null,
-      testResults: List<ViabilityTestResultModel>? = null,
-      remaining: SeedQuantityModel? = null,
-      substrate: ViabilityTestSubstrate? = null,
-      withdrawnByUserId: UserId? = null,
-      seedsCompromised: Int? = null,
-      seedsEmpty: Int? = null,
-      seedsFilled: Int? = null,
-  ): ViabilityTestModel {
-    return ViabilityTestModel(
-        accessionId = AccessionId(1),
-        id = nextViabilityTestId(),
-        remaining = remaining,
-        seedsCompromised = seedsCompromised,
-        seedsEmpty = seedsEmpty,
-        seedsFilled = seedsFilled,
-        seedsTested = seedsTested,
-        startDate = startDate,
-        substrate = substrate,
-        testResults = testResults,
-        testType = testType,
-        withdrawnByUserId = withdrawnByUserId,
-    )
-  }
-
-  private fun nextViabilityTestResultId(): ViabilityTestResultId {
-    val current = viabilityTestResultId
-    viabilityTestResultId = ViabilityTestResultId(current.value + 1)
-    return current
-  }
-
-  private fun viabilityTestResult(
-      recordingDate: LocalDate = january(2),
-      seedsGerminated: Int = 1
-  ): ViabilityTestResultModel {
-    return ViabilityTestResultModel(
-        id = nextViabilityTestResultId(),
-        recordingDate = recordingDate,
-        seedsGerminated = seedsGerminated,
-        testId = viabilityTestId)
-  }
-
-  private fun nextWithdrawalId(): WithdrawalId {
-    val current = withdrawalId
-    withdrawalId = WithdrawalId(current.value + 1)
-    return current
-  }
-
-  private fun withdrawal(
-      withdrawn: SeedQuantityModel = seeds(1),
-      date: LocalDate = january(3),
-      viabilityTestId: ViabilityTestId? = null,
-      purpose: WithdrawalPurpose =
-          if (viabilityTestId != null) WithdrawalPurpose.ViabilityTesting
-          else WithdrawalPurpose.Other,
-      remaining: SeedQuantityModel =
-          if (withdrawn.units == SeedQuantityUnits.Seeds) seeds(10) else grams(10),
-      createdTime: Instant = clock.instant(),
-      id: WithdrawalId? = nextWithdrawalId(),
-      withdrawnByUserId: UserId? = null,
-      estimatedCount: Int? =
-          if (withdrawn.units == SeedQuantityUnits.Seeds) withdrawn.quantity.toInt() else null,
-  ): WithdrawalModel {
-    return WithdrawalModel(
-        createdTime = createdTime,
-        date = date,
-        estimatedCount = estimatedCount,
-        id = id,
-        purpose = purpose,
-        remaining = remaining,
-        viabilityTestId = viabilityTestId,
-        withdrawn = withdrawn,
-        withdrawnByUserId = withdrawnByUserId,
-    )
-  }
-
-  private fun january(day: Int): LocalDate {
-    return LocalDate.of(2021, 1, day)
-  }
-
-  @Nested
-  inner class ViabilityTestCalculations {
-    @Test
-    fun `test without seeds sown is ignored`() {
-      val model =
-          accession(
-              total = seeds(10),
-              viabilityTests =
-                  listOf(
-                      viabilityTest(
-                          testResults =
-                              listOf(
-                                  viabilityTestResult(
-                                      recordingDate = january(2), seedsGerminated = 1),
-                              ),
-                      ),
-                  ),
-          )
-
-      assertNull(model.calculateLatestViabilityRecordingDate(), "Latest viability test date")
-      assertNull(model.calculateLatestViabilityPercent(), "Latest viability percent")
-      assertNull(model.calculateTotalViabilityPercent(), "Total viability percent")
-    }
-
-    @Test
-    fun `test with most recent recording date is selected`() {
-      val model =
-          accession(
-              total = seeds(100),
-              viabilityTests =
-                  listOf(
-                      viabilityTest(
-                          testType = ViabilityTestType.Lab,
-                          startDate = january(1),
-                          seedsTested = 5,
-                          testResults =
-                              listOf(
-                                  viabilityTestResult(
-                                      recordingDate = january(13), seedsGerminated = 1),
-                              ),
-                      ),
-                      viabilityTest(
-                          testType = ViabilityTestType.Lab,
-                          startDate = january(3),
-                          seedsTested = 4,
-                          testResults =
-                              listOf(
-                                  viabilityTestResult(
-                                      recordingDate = january(14), seedsGerminated = 1),
-                                  viabilityTestResult(
-                                      recordingDate = january(11), seedsGerminated = 1),
-                              ),
-                      ),
-                      viabilityTest(
-                          testType = ViabilityTestType.Lab,
-                          startDate = january(8), // most recent start date
-                          seedsTested = 3,
-                          testResults =
-                              listOf(
-                                  viabilityTestResult(
-                                      recordingDate = january(12), seedsGerminated = 3),
-                              ),
-                      ),
-                  ),
-          )
-
-      assertEquals(
-          january(14), model.calculateLatestViabilityRecordingDate(), "Latest viability test date")
-      assertEquals(50, model.calculateLatestViabilityPercent(), "Latest viability percent")
-    }
-
-    @Test
-    fun `total viability percentage includes all viability tests and cut test`() {
-      val model =
-          accession(
-              total = seeds(30),
-              viabilityTests =
-                  listOf(
-                      viabilityTest(
-                          seedsTested = 10,
-                          testResults =
-                              listOf(
-                                  viabilityTestResult(
-                                      recordingDate = january(2), seedsGerminated = 4),
-                                  viabilityTestResult(
-                                      recordingDate = january(3), seedsGerminated = 3),
-                              ),
-                      ),
-                      viabilityTest(
-                          startDate = january(8),
-                          seedsTested = 5,
-                          testResults =
-                              listOf(
-                                  viabilityTestResult(
-                                      recordingDate = january(9), seedsGerminated = 1),
-                              ),
-                      ),
-                  ),
-              cutTestSeedsCompromised = 2,
-              cutTestSeedsEmpty = 1,
-              cutTestSeedsFilled = 5,
-          )
-
-      // 13 seeds viable of 23 tested
-      val expectedPercentage = 56
-
-      assertEquals(expectedPercentage, model.calculateTotalViabilityPercent())
-    }
-
-    @Test
-    fun `total viability percentage not set if no test results`() {
-      val model =
-          accession(
-              total = seeds(20),
-              viabilityTests = listOf(viabilityTest(seedsTested = 10)),
-          )
-
-      assertNull(model.calculateTotalViabilityPercent())
-    }
-
-    @Test
-    fun `total viability percentage ignores cut test results if values are missing`() {
-      val missingEmpty =
-          accession(
-              total = seeds(50),
-              viabilityTests =
-                  listOf(
-                      viabilityTest(
-                          seedsTested = 10,
-                          testResults =
-                              listOf(
-                                  viabilityTestResult(
-                                      recordingDate = january(2), seedsGerminated = 4),
-                              ),
-                      ),
-                  ),
-              cutTestSeedsCompromised = 2,
-              cutTestSeedsFilled = 5,
-          )
-
-      val missingCompromised =
-          missingEmpty.copy(cutTestSeedsEmpty = 1, cutTestSeedsCompromised = null)
-      val missingFilled = missingEmpty.copy(cutTestSeedsEmpty = 1, cutTestSeedsFilled = null)
-
-      // 4 seeds germinated out of 10 sown; cut test is ignored
-      val expectedPercentage = 40
-
-      assertEquals(
-          expectedPercentage, missingEmpty.calculateTotalViabilityPercent(), "Empty count missing")
-      assertEquals(
-          expectedPercentage,
-          missingCompromised.calculateTotalViabilityPercent(),
-          "Compromised count missing")
-      assertEquals(
-          expectedPercentage,
-          missingFilled.calculateTotalViabilityPercent(),
-          "Filled count missing")
-    }
-
-    @Test
-    fun `most recent recording date is used as test date`() {
-      val model =
-          accession(
-              total = seeds(10),
-              viabilityTests =
-                  listOf(
-                      viabilityTest(
-                          seedsTested = 1,
-                          testResults =
-                              listOf(
-                                  viabilityTestResult(
-                                      recordingDate = january(3), seedsGerminated = 1),
-                                  viabilityTestResult(
-                                      recordingDate = january(4), seedsGerminated = 2),
-                                  viabilityTestResult(
-                                      recordingDate = january(2), seedsGerminated = 3),
-                              ),
-                      ),
-                  ),
-          )
-
-      assertEquals(january(4), model.calculateLatestViabilityRecordingDate())
-    }
-
-    @Test
-    fun `viability percent is not auto-populated on v2 accessions when test results are added`() {
-      val model =
-          accession()
-              .copy(isManualState = true, remaining = seeds(50))
-              .withCalculatedValues(clock)
-              .addViabilityTest(
-                  viabilityTest(
-                      seedsTested = 10,
-                      testResults =
-                          listOf(
-                              viabilityTestResult(
-                                  recordingDate = january(2), seedsGerminated = 4))),
-                  clock)
-
-      assertNull(model.totalViabilityPercent)
-    }
-
-    @Test
-    fun `viability percent is not cleared on v2 accessions without tests`() {
-      val percent = 77
-      val model =
-          accession()
-              .copy(isManualState = true, remaining = seeds(50), totalViabilityPercent = percent)
-              .withCalculatedValues(clock)
-
-      assertEquals(percent, model.totalViabilityPercent)
-    }
-
-    @Test
-    fun `viability percent is not overwritten on v2 accessions when test results are added`() {
-      val percent = 16
-      val model =
-          accession()
-              .copy(isManualState = true, remaining = seeds(50), totalViabilityPercent = percent)
-              .withCalculatedValues(clock)
-              .addViabilityTest(
-                  viabilityTest(
-                      seedsTested = 5,
-                      testResults =
-                          listOf(
-                              viabilityTestResult(
-                                  recordingDate = january(2), seedsGerminated = 5))),
-                  clock)
-
-      assertEquals(percent, model.totalViabilityPercent)
-    }
-
-    @Test
-    fun `viability test withdrawnByUserId is propagated to new withdrawals`() {
-      val withdrawnByUserId = UserId(1234)
-      val model =
-          accession()
-              .copy(isManualState = true, remaining = seeds(10))
-              .withCalculatedValues(clock)
-              .addViabilityTest(
-                  viabilityTest(seedsTested = 1, withdrawnByUserId = withdrawnByUserId), clock)
-
-      assertEquals(withdrawnByUserId, model.withdrawals[0].withdrawnByUserId)
-    }
-
-    @Test
-    fun `viability test withdrawnByUserId change is propagated to existing withdrawals`() {
-      val oldWithdrawnByUserId = UserId(1234)
-      val newWithdrawnByUserId = UserId(5678)
-
-      val viabilityTest = viabilityTest(seedsTested = 1, withdrawnByUserId = oldWithdrawnByUserId)
-      val viabilityTestId = viabilityTest.id!!
-
-      val model =
-          accession()
-              .copy(
-                  isManualState = true,
-                  remaining = seeds(10),
-                  viabilityTests = listOf(viabilityTest),
-                  withdrawals =
-                      listOf(
-                          withdrawal(
-                              purpose = WithdrawalPurpose.ViabilityTesting,
-                              viabilityTestId = viabilityTestId)))
-              .withCalculatedValues(clock)
-              .updateViabilityTest(viabilityTestId, clock) {
-                it.copy(withdrawnByUserId = newWithdrawnByUserId)
-              }
-
-      assertEquals(newWithdrawnByUserId, model.withdrawals[0].withdrawnByUserId)
-    }
-
-    @Test
-    fun `viability test update with null withdrawnByUserId does not change user ID of existing withdrawal`() {
-      val withdrawnByUserId = UserId(1234)
-
-      val viabilityTest = viabilityTest(seedsTested = 1, withdrawnByUserId = withdrawnByUserId)
-      val viabilityTestId = viabilityTest.id!!
-
-      val model =
-          accession()
-              .copy(
-                  isManualState = true,
-                  remaining = seeds(10),
-                  viabilityTests = listOf(viabilityTest),
-                  withdrawals =
-                      listOf(
-                          withdrawal(
-                              purpose = WithdrawalPurpose.ViabilityTesting,
-                              viabilityTestId = viabilityTestId)))
-              .withCalculatedValues(clock)
-              .updateViabilityTest(viabilityTestId, clock) { it.copy(withdrawnByUserId = null) }
-
-      assertEquals(withdrawnByUserId, model.withdrawals[0].withdrawnByUserId)
-    }
-
-    @Test
-    fun `cut test results are reflected in v1 cut test fields`() {
-      val model =
-          accession()
-              .copy(isManualState = true, remaining = seeds(10))
-              .addViabilityTest(
-                  viabilityTest(
-                      ViabilityTestType.Cut,
-                      seedsCompromised = 1,
-                      seedsEmpty = 2,
-                      seedsFilled = 3,
-                      seedsTested = 6),
-                  clock)
-              .addViabilityTest(
-                  viabilityTest(ViabilityTestType.Cut, seedsFilled = 1, seedsTested = 1), clock)
-
-      assertEquals(1, model.cutTestSeedsCompromised, "Compromised")
-      assertEquals(2, model.cutTestSeedsEmpty, "Empty")
-      assertEquals(4, model.cutTestSeedsFilled, "Filled")
-    }
-
-    @Test
-    fun `new cut test is added if v1 cut test fields are newly populated`() {
-      val initial = accession(total = seeds(10))
-      val updated =
-          initial
-              .copy(cutTestSeedsCompromised = 1, cutTestSeedsEmpty = 2, cutTestSeedsFilled = 3)
-              .withCalculatedValues(clock, initial)
-
-      val expectedTest =
-          ViabilityTestModel(
-              remaining = seeds(4),
-              seedsCompromised = 1,
-              seedsEmpty = 2,
-              seedsFilled = 3,
-              seedsTested = 6,
-              testType = ViabilityTestType.Cut,
-          )
-
-      assertEquals(listOf(expectedTest), updated.viabilityTests)
-    }
-
-    @Test
-    fun `cut test is removed if v1 cut test fields are cleared`() {
-      val initial = accession(total = seeds(10)).withCalculatedValues(clock)
-      val withCutTest = initial.copy(cutTestSeedsEmpty = 1).withCalculatedValues(clock, initial)
-      val updated =
-          withCutTest.copy(cutTestSeedsEmpty = null).withCalculatedValues(clock, withCutTest)
-
-      assertNotEquals(emptyList<ViabilityTestModel>(), withCutTest.viabilityTests, "Before edit")
-      assertEquals(emptyList<ViabilityTestModel>(), updated.viabilityTests, "After edit")
-    }
-
-    @Test
-    fun `existing cut test is updated if v1 cut test fields are modified`() {
-      val initialViabilityTest =
-          ViabilityTestModel(
-              id = ViabilityTestId(1),
-              remaining = seeds(9),
-              seedsFilled = 1,
-              seedsTested = 1,
-              testType = ViabilityTestType.Cut)
-      val initialWithdrawal =
-          withdrawal(
-              remaining = seeds(9), viabilityTestId = initialViabilityTest.id, withdrawn = seeds(1))
-
-      val initialAccession =
-          accession(
-                  cutTestSeedsFilled = 1,
-                  total = seeds(10),
-                  viabilityTests = listOf(initialViabilityTest),
-                  withdrawals = listOf(initialWithdrawal))
-              .withCalculatedValues(clock)
-      val updatedAccession =
-          initialAccession
-              .copy(
-                  cutTestSeedsCompromised = 1,
-                  cutTestSeedsFilled = 2,
-                  // v1 PUT requests won't include the cut test or its withdrawal since we filter
-                  // them out of v1 GET responses.
-                  viabilityTests = emptyList(),
-                  withdrawals = emptyList())
-              .withCalculatedValues(clock, initialAccession)
-
-      val expectedViabilityTest =
-          initialViabilityTest.copy(
-              remaining = seeds(7), seedsCompromised = 1, seedsFilled = 2, seedsTested = 3)
-      val expectedWithdrawal =
-          initialWithdrawal.copy(
-              estimatedCount = 3,
-              remaining = seeds(7),
-              viabilityTest = expectedViabilityTest,
-              withdrawn = seeds(3))
-
-      assertEquals(
-          listOf(expectedViabilityTest),
-          updatedAccession.viabilityTests,
-          "Existing viability test should be updated")
-      assertEquals(
-          listOf(expectedWithdrawal),
-          updatedAccession.withdrawals,
-          "Existing withdrawal should be updated")
-      assertEquals(seeds(7), updatedAccession.remaining, "Remaining quantity")
-    }
-
-    @Test
-    fun `cut test on weight-based accession requires subset data`() {
-      val initial = accession(total = grams(10)).withCalculatedValues(clock)
-
-      assertThrows<IllegalArgumentException> {
-        initial.copy(cutTestSeedsFilled = 1).withCalculatedValues(clock, initial)
-      }
-    }
-  }
-
-  @Suppress("UnusedDataClassCopyResult")
-  @Nested
-  inner class ValidationRules {
-    @Test
-    fun `cannot specify accession size in seeds if processing method is Weight`() {
-      assertThrows<IllegalArgumentException> {
-        accession(processingMethod = ProcessingMethod.Weight, total = seeds(1))
-      }
-    }
-
-    @Test
-    fun `cannot specify accession size in weight if processing method is Count`() {
-      assertThrows<IllegalArgumentException> {
-        accession(processingMethod = ProcessingMethod.Count, total = grams(1))
-      }
-    }
-
-    @Test
-    fun `processing method is required if total size specified`() {
-      assertThrows<IllegalArgumentException> {
-        accession(processingMethod = null, total = grams(1))
-      }
-    }
-
-    @Test
-    fun `cannot withdraw more seeds than exist in the accession`() {
-      assertThrows<IllegalArgumentException> {
-        accession(
-            processingMethod = ProcessingMethod.Count,
-            total = seeds(1),
-            viabilityTests = listOf(viabilityTest(seedsTested = 1)),
-            withdrawals = listOf(withdrawal(withdrawn = seeds(1), date = tomorrow)))
-      }
-    }
-
-    @Test
-    fun `cannot add withdrawal with more seeds than exist in the accession`() {
-      val accession =
-          accession().copy(isManualState = true, remaining = seeds(10)).withCalculatedValues(clock)
-
-      assertThrows<IllegalArgumentException> {
-        accession.addWithdrawal(WithdrawalModel(date = today, withdrawn = seeds(11)), tomorrowClock)
-      }
-    }
-
-    @Test
-    fun `cannot specify negative seeds remaining for weight-based accessions`() {
-      assertThrows<IllegalArgumentException>("Viability tests") {
-        accession(
-            processingMethod = ProcessingMethod.Weight,
-            total = grams(1),
-            viabilityTests = listOf(viabilityTest(seedsTested = 1, remaining = grams(-1))))
-      }
-
-      assertThrows<IllegalArgumentException>("withdrawals") {
-        accession(
-            processingMethod = ProcessingMethod.Weight,
-            total = grams(1),
-            withdrawals = listOf(withdrawal(withdrawn = grams(1), remaining = grams(-1))))
-      }
-    }
-
-    @Test
-    fun `total quantity must be greater than zero`() {
-      assertThrows<IllegalArgumentException>("zero seeds") {
-        accession(processingMethod = ProcessingMethod.Count, total = seeds(0))
-      }
-      assertThrows<IllegalArgumentException>("zero weight") {
-        accession(processingMethod = ProcessingMethod.Weight, total = grams(0))
-      }
-      assertThrows<IllegalArgumentException>("negative seeds") {
-        accession(processingMethod = ProcessingMethod.Count, total = seeds(-1))
-      }
-      assertThrows<IllegalArgumentException>("negative weight") {
-        accession(processingMethod = ProcessingMethod.Weight, total = grams(-1))
-      }
-    }
-
-    @Test
-    fun `cannot withdraw seeds if remaining quantity is not set`() {
-      val initial = accession().copy(isManualState = true)
-
-      assertThrows<IllegalArgumentException> {
-        initial.copy(withdrawals = listOf(withdrawal(withdrawn = seeds(1))))
-      }
-    }
-
-    @Test
-    fun `cannot withdraw more seeds than most recent observed quantity`() {
-      val initial =
-          accession().copy(isManualState = true, remaining = seeds(10)).withCalculatedValues(clock)
-
-      assertThrows<IllegalArgumentException> {
-        initial
-            .copy(withdrawals = listOf(withdrawal(withdrawn = seeds(11))))
-            .withCalculatedValues(clock, initial)
-      }
-    }
-
-    @Test
-    fun `cannot withdraw more weight than most recent observed seed count`() {
-      val initial =
-          accession()
-              .copy(
-                  isManualState = true,
-                  subsetCount = 2,
-                  subsetWeightQuantity = grams(2),
-                  remaining = seeds(10))
-              .withCalculatedValues(clock)
-
-      assertThrows<IllegalArgumentException> {
-        initial
-            .copy(withdrawals = listOf(withdrawal(withdrawn = grams(11))))
-            .withCalculatedValues(clock, initial)
-      }
-    }
-
-    @Test
-    fun `cannot remove remaining quantity once it has been set`() {
-      val initial =
-          accession().copy(isManualState = true, remaining = seeds(1)).withCalculatedValues(clock)
-
-      assertThrows<IllegalArgumentException> { initial.copy(remaining = null) }
-    }
-
-    @Test
-    fun `cannot change remaining quantity from seeds to weight without subset info if withdrawals exist`() {
-      val initial =
-          accession()
-              .copy(isManualState = true, remaining = seeds(10))
-              .withCalculatedValues(clock)
-              .copy(withdrawals = listOf(withdrawal(seeds(1))))
-              .withCalculatedValues(clock)
-
-      assertThrows<IllegalArgumentException> { initial.copy(remaining = grams(1)) }
-    }
-
-    @Test
-    fun `cannot change remaining quantity from weight to seeds without subset info if withdrawals exist`() {
-      val initial =
-          accession()
-              .copy(isManualState = true, remaining = grams(10))
-              .withCalculatedValues(clock)
-              .copy(withdrawals = listOf(withdrawal(grams(1))))
-              .withCalculatedValues(clock)
-
-      assertThrows<IllegalArgumentException> { initial.copy(remaining = seeds(1)) }
-    }
-
-    @Test
-    fun `can change remaining quantity from seeds to weight without subset info if no withdrawals exist`() {
-      val initial =
-          accession().copy(isManualState = true, remaining = seeds(10)).withCalculatedValues(clock)
-
-      assertDoesNotThrow { initial.copy(remaining = grams(1)) }
-    }
-
-    @Test
-    fun `can change remaining quantity from weight to seeds without subset info if no withdrawals exist`() {
-      val initial =
-          accession().copy(isManualState = true, remaining = grams(10)).withCalculatedValues(clock)
-
-      assertDoesNotThrow { initial.copy(remaining = seeds(1)) }
-    }
-
-    @Test
-    fun `can change remaining quantity from seeds to weight if subset info exists`() {
-      val initial =
-          accession()
-              .copy(
-                  isManualState = true,
-                  subsetCount = 1,
-                  subsetWeightQuantity = grams(1),
-                  remaining = seeds(10))
-              .withCalculatedValues(clock)
-              .copy(withdrawals = listOf(withdrawal(seeds(1))))
-              .withCalculatedValues(clock)
-
-      assertDoesNotThrow { initial.copy(remaining = grams(10)) }
-    }
-
-    @Test
-    fun `can change remaining quantity from weight to seeds if subset info exists`() {
-      val initial =
-          accession()
-              .copy(
-                  isManualState = true,
-                  subsetCount = 1,
-                  subsetWeightQuantity = grams(1),
-                  remaining = grams(10))
-              .withCalculatedValues(clock)
-              .copy(withdrawals = listOf(withdrawal(grams(1))))
-              .withCalculatedValues(clock)
-
-      assertDoesNotThrow { initial.copy(remaining = seeds(10)) }
-    }
-
-    @Test
-    fun `cannot create viability test on weight-based accession without subset data`() {
-      val initial =
-          accession().copy(isManualState = true, remaining = grams(10)).withCalculatedValues(clock)
-
-      assertThrows<IllegalArgumentException> {
-        initial.addViabilityTest(viabilityTest(seedsTested = 1, startDate = null), tomorrowClock)
-      }
-    }
-
-    @Test
-    fun `cannot add viability test with more seeds sown than remaining quantity`() {
-      val initial =
-          accession().copy(isManualState = true, remaining = seeds(10)).withCalculatedValues(clock)
-
-      assertThrows<IllegalArgumentException> {
-        initial.addViabilityTest(viabilityTest(seedsTested = 11, startDate = null), tomorrowClock)
-      }
-    }
-
-    @Test
-    fun `can add viability test with more seeds sown than remaining quantity if it is older than the latest observed quantity`() {
-      val initial =
-          accession().copy(isManualState = true, remaining = seeds(10)).withCalculatedValues(clock)
-
-      assertDoesNotThrow {
-        initial.addViabilityTest(
-            viabilityTest(seedsTested = 11, startDate = yesterday), tomorrowClock)
-      }
-    }
-
-    @Test
-    fun `can add viability test with fewer seeds sown than estimated seed count by weight`() {
-      val initial =
-          accession()
-              .copy(
-                  isManualState = true,
-                  subsetCount = 2,
-                  subsetWeightQuantity = grams(2),
-                  remaining = grams(10))
-              .withCalculatedValues(clock)
-
-      assertDoesNotThrow { initial.addViabilityTest(viabilityTest(seedsTested = 9), tomorrowClock) }
-    }
-
-    @Test
-    fun `cannot add new viability test with more seeds sown than estimated seed count by weight`() {
-      val initial =
-          accession()
-              .copy(
-                  isManualState = true,
-                  subsetCount = 2,
-                  subsetWeightQuantity = grams(2),
-                  remaining = grams(10))
-              .withCalculatedValues(clock)
-
-      assertThrows<IllegalArgumentException> {
-        initial.addViabilityTest(viabilityTest(seedsTested = 11, startDate = null), tomorrowClock)
-      }
-    }
-
-    @Test
-    fun `cannot add new viability test with substrate that is not valid for test type`() {
-      val initial =
-          accession().copy(isManualState = true, remaining = seeds(10)).withCalculatedValues(clock)
-
-      assertThrows<IllegalArgumentException> {
-        initial.addViabilityTest(
-            viabilityTest(
-                seedsTested = 1,
-                substrate = ViabilityTestSubstrate.Paper,
-                testType = ViabilityTestType.Nursery),
-            clock)
-      }
-    }
-  }
-
+internal class AccessionModelCalculationsTest {
   @Nested
   inner class WeightBasedAccessionCalculations {
     @Test
@@ -1980,151 +1138,6 @@ internal class AccessionModelTest {
               .withCalculatedValues(clock)
 
       assertEquals(grams(5), accession.estimatedWeight)
-    }
-  }
-
-  @Nested
-  inner class V1ToV2Conversion {
-    @Test
-    fun `weight-based accession with no subset data and a viability test`() {
-      val initial =
-          AccessionModel(
-              checkedInTime = yesterdayInstant,
-              createdTime = yesterdayInstant,
-              processingMethod = ProcessingMethod.Weight,
-              total = grams(10),
-              viabilityTests =
-                  listOf(
-                      ViabilityTestModel(
-                          id = ViabilityTestId(1),
-                          remaining = grams(9),
-                          seedsTested = 1,
-                          startDate = today,
-                          testType = ViabilityTestType.Lab,
-                      ),
-                  ),
-              withdrawals =
-                  listOf(
-                      WithdrawalModel(
-                          createdTime = todayInstant,
-                          date = today,
-                          id = WithdrawalId(1),
-                          purpose = WithdrawalPurpose.ViabilityTesting,
-                          remaining = grams(9),
-                          viabilityTestId = ViabilityTestId(1),
-                          withdrawn = seeds(1),
-                      ),
-                  ),
-          )
-
-      val expected =
-          AccessionModel(
-              checkedInTime = yesterdayInstant,
-              createdTime = yesterdayInstant,
-              estimatedWeight = grams(9),
-              isManualState = true,
-              latestObservedQuantity = grams(9),
-              latestObservedTime = todayInstant,
-              processingMethod = ProcessingMethod.Weight,
-              remaining = grams(9),
-              state = AccessionState.AwaitingProcessing,
-              total = grams(10),
-              viabilityTests =
-                  listOf(
-                      ViabilityTestModel(
-                          id = ViabilityTestId(1),
-                          remaining = grams(9),
-                          seedsTested = 1,
-                          startDate = today,
-                          testType = ViabilityTestType.Lab,
-                      ),
-                  ),
-              withdrawals =
-                  listOf(
-                      WithdrawalModel(
-                          createdTime = todayInstant,
-                          date = today,
-                          estimatedCount = 1,
-                          estimatedWeight = grams(1),
-                          id = WithdrawalId(1),
-                          purpose = WithdrawalPurpose.ViabilityTesting,
-                          remaining = grams(9),
-                          viabilityTest =
-                              ViabilityTestModel(
-                                  id = ViabilityTestId(1),
-                                  remaining = grams(9),
-                                  seedsTested = 1,
-                                  startDate = today,
-                                  testType = ViabilityTestType.Lab,
-                              ),
-                          viabilityTestId = ViabilityTestId(1),
-                          weightDifference = grams(1),
-                          withdrawn = seeds(1),
-                      ),
-                  ),
-          )
-
-      assertJsonEquals(expected, initial.toV2Compatible(tomorrowClock))
-    }
-  }
-
-  @Nested
-  inner class V2ToV1Conversion {
-    @Test
-    fun `viability percent is overwritten with computed value`() {
-      val expectedPercent = 40
-      val v2Model =
-          accession()
-              .copy(isManualState = true, remaining = seeds(500))
-              .withCalculatedValues(clock)
-              .addViabilityTest(
-                  viabilityTest(
-                      seedsTested = 100,
-                      testResults =
-                          listOf(
-                              viabilityTestResult(
-                                  recordingDate = january(2), seedsGerminated = expectedPercent))),
-                  clock)
-      assertNull(v2Model.totalViabilityPercent)
-
-      val v1Model = v2Model.toV1Compatible(clock)
-      assertEquals(40, v1Model.totalViabilityPercent)
-    }
-
-    @Test
-    fun `backdated viability tests with more seeds than initial quantity`() {
-      val initialV2Model =
-          accession()
-              .copy(isManualState = true, remaining = seeds(2))
-              .withCalculatedValues(yesterdayClock)
-      val v2Model =
-          initialV2Model
-              .copy(latestObservedQuantityCalculated = false, remaining = seeds(1))
-              .withCalculatedValues(clock, initialV2Model)
-              // In v2, this is valid because it is dated yesterday and we have an observed quantity
-              // from today that overrides what would otherwise be a negative seeds remaining value.
-              .addViabilityTest(viabilityTest(seedsTested = 3, startDate = yesterday), clock)
-
-      val v1Model = v2Model.toV1Compatible(clock)
-      assertEquals(seeds(1), v1Model.viabilityTests.getOrNull(0)?.remaining)
-    }
-
-    @Test
-    fun `backdated withdrawals with more seeds than initial quantity`() {
-      val initialV2Model =
-          accession()
-              .copy(isManualState = true, remaining = seeds(2))
-              .withCalculatedValues(yesterdayClock)
-      val v2Model =
-          initialV2Model
-              .copy(latestObservedQuantityCalculated = false, remaining = seeds(1))
-              .withCalculatedValues(clock, initialV2Model)
-              // In v2, this is valid because it is yesterday and we have an observed quantity from
-              // today that overrides what would otherwise be a negative seeds remaining value.
-              .addWithdrawal(withdrawal(seeds(3), date = yesterday, id = null), clock)
-
-      val v1Model = v2Model.toV1Compatible(clock)
-      assertEquals(seeds(1), v1Model.withdrawals.getOrNull(0)?.remaining)
     }
   }
 }
